attaching and removing event listener can be reused
