+++ 12-Use-popcorn/src/App.js
@@ -11,6 +11,7 @@ import WatchedSummary from './presentational/WatchedSummary'
 import WatchedMoviesList from './presentational/WatchedMoviesList'
 import { useMovies } from './useMovies'
 import { useLocalStorageState } from './useLocalStorageState'
+import { useKey } from './useKey'
 
 /* const tempMovieData = [
   {
@@ -178,23 +179,7 @@ function MovieDetails({ selectedId, onCloseMovie, onAddWatched, watched }) {
     [title]
   )
 
-  useEffect(
-    function () {
-      function callback(e) {
-        if (e.code === 'Escape') {
-          onCloseMovie()
-          console.log('closing')
-        }
-      }
-
-      document.addEventListener('keydown', callback)
-
-      return function () {
-        document.removeEventListener('keydown', callback)
-      }
-    },
-    [onCloseMovie]
-  )
+  useKey('Escape', onCloseMovie)
 
   useEffect(
     function () {
